refactor(frontend): split rpc-api.ts into domain modules
