+++ .changesets/1782956392-refactor-frontend-split-rpc-api-into-domain-modules-lo5p.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(frontend): split rpc-api into domain modules
